글 삭제 인가 작업

+++ src/main/java/com/back/domain/post/post/controller/ApiV1PostController.java
@@ -135,7 +135,7 @@ public class ApiV1PostController {
         Post post = postService.findById(id).get();
 
         if(!actor.equals(post.getAuthor())){
-            throw new ServiceException("403-1","권한이 없습니다.");
+            throw new ServiceException("403-1","수정 권한이 없습니다.");
         }
 
         post = postService.modify(id, reqBody.title, reqBody.content);
@@ -152,8 +152,18 @@ public class ApiV1PostController {
     @DeleteMapping("/{id}")
     @Operation(summary="글 삭제")
     public RsData<Void> delete(
-            @PathVariable int id
+            @PathVariable int id,
+            @RequestHeader("Authorization") String apiKey
     ) {
+        apiKey = apiKey.replace("Bearer ", ""); // "Bearer " 접두어 제거
+        Member actor = memberService.findByApiKey(apiKey).orElseThrow(
+                () -> new ServiceException("401-1","유효하지 않은 API Key 입니다.")
+        ); //인증
+
+        Post post = postService.findById(id).get();
+        if(!actor.equals(post.getAuthor())){
+            throw new ServiceException("403-1","삭제 권한이 없습니다.");
+        }
 
         postService.deleteById(id);
